requirements.txt
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
dnspython==2.2.1
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==6.0.0
install==1.3.5
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
pycparser==2.21
PyMySQL==1.0.2
python-decouple==3.8
python-http-client==3.3.7
sendgrid==6.10.0
SQLAlchemy==1.4.44
starkbank-ecdsa==2.2.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
wtforms-validators==1.0.0
zipp==3.11.0
<<<<<<< HEAD
Flask-CKEditor==0.4.6

pandas~=2.0.1
=======
>>>>>>> 3265474 (fix some bugs, remove credential and add env)
